--- a/src/Entity/OptimizableData.hs
+++ b/src/Entity/OptimizableData.hs
@@ -3,5 +3,4 @@ module Entity.OptimizableData (OptimizableData (..)) where
 data OptimizableData
   = Enum
   | Unary -- for newtype-ish optimization
-  | Single
   deriving (Show)