--- a/src/Scene/Parse/Discern/Fallback.hs
+++ b/src/Scene/Parse/Discern/Fallback.hs
@@ -5,6 +5,7 @@ import Context.Gensym qualified as Gensym
 import Context.Throw qualified as Throw
 import Control.Comonad.Cofree
 import Data.Vector qualified as V
+import Entity.Binder
 import Entity.Ident
 import Entity.Noema qualified as N
 import Entity.Pattern
@@ -15,26 +16,28 @@ import Scene.Parse.Discern.Noema
 getFallbackMatrix ::
   N.IsNoetic ->
   Ident ->
-  PatternMatrix ([Ident], WT.WeakTerm) ->
-  App (PatternMatrix ([Ident], WT.WeakTerm))
+  PatternMatrix ([Ident], [(BinderF WT.WeakTerm, WT.WeakTerm)], WT.WeakTerm) ->
+  App (PatternMatrix ([Ident], [(BinderF WT.WeakTerm, WT.WeakTerm)], WT.WeakTerm))
 getFallbackMatrix isNoetic cursor mat = do
   mapMaybeRowM (fallbackRow isNoetic cursor) mat
 
 fallbackRow ::
   N.IsNoetic ->
   Ident ->
-  PatternRow ([Ident], WT.WeakTerm) ->
-  App (Maybe (PatternRow ([Ident], WT.WeakTerm)))
-fallbackRow isNoetic cursor (patternVector, (freedVars, body@(mBody :< _))) =
+  PatternRow ([Ident], [(BinderF WT.WeakTerm, WT.WeakTerm)], WT.WeakTerm) ->
+  App (Maybe (PatternRow ([Ident], [(BinderF WT.WeakTerm, WT.WeakTerm)], WT.WeakTerm)))
+fallbackRow isNoetic cursor (patternVector, (freedVars, baseSeq, body@(mBody :< _))) =
   case V.uncons patternVector of
     Nothing ->
       Throw.raiseCritical' "defaulting against the empty pattern matrix shouldn't happen"
     Just ((_, WildcardVar), rest) ->
-      return $ Just (rest, (freedVars, body))
+      return $ Just (rest, (freedVars, baseSeq, body))
     Just ((_, Var x), rest) -> do
       h <- Gensym.newHole mBody []
       adjustedCursor <- castToNoemaIfNecessary isNoetic (mBody :< WT.Var cursor)
-      let body' = mBody :< WT.Let WT.Clear (mBody, x, h) adjustedCursor body
-      return $ Just (rest, (freedVars, body'))
+      -- let body' = mBody :< WT.Let WT.Opaque (mBody, x, h) adjustedCursor body
+      -- return $ Just (rest, (freedVars,  body'))
+      -- let body' = mBody :< WT.Let WT.Opaque (mBody, x, h) adjustedCursor body
+      return $ Just (rest, (freedVars, ((mBody, x, h), adjustedCursor) : baseSeq, body))
     Just ((_, Cons {}), _) ->
       return Nothing