--- a/src/Scene/Parse/Discern/Specialize.hs
+++ b/src/Scene/Parse/Discern/Specialize.hs
@@ -7,6 +7,7 @@ import Context.Throw qualified as Throw
 import Control.Comonad.Cofree
 import Data.Vector qualified as V
 import Entity.ArgNum qualified as AN
+import Entity.Binder
 import Entity.DefiniteDescription qualified as DD
 import Entity.Ident
 import Entity.Noema qualified as N
@@ -20,8 +21,8 @@ specialize ::
   N.IsNoetic ->
   Ident ->
   (DD.DefiniteDescription, AN.ArgNum) ->
-  PatternMatrix ([Ident], WT.WeakTerm) ->
-  App (PatternMatrix ([Ident], WT.WeakTerm))
+  PatternMatrix ([Ident], [(BinderF WT.WeakTerm, WT.WeakTerm)], WT.WeakTerm) ->
+  App (PatternMatrix ([Ident], [(BinderF WT.WeakTerm, WT.WeakTerm)], WT.WeakTerm))
 specialize isNoetic cursor cons mat = do
   mapMaybeRowM (specializeRow isNoetic cursor cons) mat
 
@@ -29,30 +30,29 @@ specializeRow ::
   N.IsNoetic ->
   Ident ->
   (DD.DefiniteDescription, AN.ArgNum) ->
-  PatternRow ([Ident], WT.WeakTerm) ->
-  App (Maybe (PatternRow ([Ident], WT.WeakTerm)))
-specializeRow isNoetic cursor (dd, argNum) (patternVector, (freedVars, body@(mBody :< _))) =
+  PatternRow ([Ident], [(BinderF WT.WeakTerm, WT.WeakTerm)], WT.WeakTerm) ->
+  App (Maybe (PatternRow ([Ident], [(BinderF WT.WeakTerm, WT.WeakTerm)], WT.WeakTerm)))
+specializeRow isNoetic cursor (dd, argNum) (patternVector, (freedVars, baseSeq, body@(mBody :< _))) =
   case V.uncons patternVector of
     Nothing ->
       Throw.raiseCritical' "specialization against the empty pattern matrix shouldn't happen"
     Just ((m, WildcardVar), rest) -> do
       let wildcards = V.fromList $ replicate (AN.reify argNum) (m, WildcardVar)
-      return $ Just (V.concat [wildcards, rest], (freedVars, body))
+      return $ Just (V.concat [wildcards, rest], (freedVars, baseSeq, body))
     Just ((_, Var x), rest) -> do
       let wildcards = V.fromList $ replicate (AN.reify argNum) (mBody, WildcardVar)
       h <- Gensym.newHole mBody []
       adjustedCursor <- castToNoemaIfNecessary isNoetic (mBody :< WT.Var cursor)
-      let body' = mBody :< WT.Let WT.Clear (mBody, x, h) adjustedCursor body
-      return $ Just (V.concat [wildcards, rest], (freedVars, body'))
+      return $ Just (V.concat [wildcards, rest], (freedVars, ((mBody, x, h), adjustedCursor) : baseSeq, body))
     Just ((_, Cons (ConsInfo {..})), rest) ->
       if dd == consDD
         then do
           od <- OptimizableData.lookup consDD
           case od of
             Just OD.Enum ->
-              return $ Just (V.concat [V.fromList args, rest], (freedVars, body))
+              return $ Just (V.concat [V.fromList args, rest], (freedVars, baseSeq, body))
             Just OD.Unary ->
-              return $ Just (V.concat [V.fromList args, rest], (freedVars, body))
+              return $ Just (V.concat [V.fromList args, rest], (freedVars, baseSeq, body))
             _ ->
-              return $ Just (V.concat [V.fromList args, rest], (cursor : freedVars, body))
+              return $ Just (V.concat [V.fromList args, rest], (cursor : freedVars, baseSeq, body))
         else return Nothing