--- a/test/statement/resource-basic/expected
+++ b/test/statement/resource-basic/expected
@@ -1,5 +1,1 @@
-copied.
-copied.
-discarded.
-discarded.
 discarded.