… src/Scene/Term/Subst.hs
… src/Scene/WeakTerm/Fill.hs
… src/Scene/WeakTerm/Reduce.hs
… src/Scene/WeakTerm/Subst.hs
… test/misc/args/module.ens
… test/misc/calc/module.ens
… test/misc/check/module.ens
… test/misc/codata-basic/module.ens
… test/misc/complex-cancel/module.ens
… test/misc/ex-falso/module.ens
… test/misc/fact/module.ens
… test/misc/file/module.ens
… test/misc/file/source/file.nt
… test/misc/fix-and-free-vars/module.ens
… test/misc/fold-tails/module.ens
… test/misc/foreign/module.ens
… test/misc/hello/module.ens
… test/misc/lambda-list/module.ens
… test/misc/mutable/module.ens
… test/misc/nat-fact/module.ens
… test/misc/nat-list/module.ens
… test/misc/print-float/module.ens
… test/pfds/binary-search-tree/module.ens
… test/pfds/binomial-heap/module.ens
… test/pfds/custom-stack/module.ens
… test/pfds/finite-map/module.ens
… test/pfds/leftist-heap/module.ens
… test/pfds/naive-queue/module.ens
… test/pfds/pairing-heap/module.ens
… test/pfds/random-access-list/module.ens