before (510e6aeb):
```
❯ hyperfine "neut clean && neut build"
Benchmark 1: neut clean && neut build
  Time (mean ± σ):      1.855 s ±  0.015 s    [User: 2.977 s, System: 0.710 s]
  Range (min … max):    1.837 s …  1.882 s    10 runs
```

after:
```
❯ hyperfine "neut clean && neut build"
Benchmark 1: neut clean && neut build
  Time (mean ± σ):      1.283 s ±  0.011 s    [User: 2.238 s, System: 0.616 s]
  Range (min … max):    1.263 s …  1.298 s    10 runs
```

(tested in vekatze/neut-core@403af49)